Fixes Merchant's "Vegetables Pack" in the pantry. (#3057)
## About The Pull Request

Previously, the "Vegetables Pack" purchaseable by the merchant gave the
wrong items for cabbage, onion, and potato. These were basically
doppelganger versions of their respective items that couldn't be used in
recipes and couldn't be cut, because they had the path
/obj/item/reagent_containers/food/snacks/grown/cabbage instead of
/obj/item/reagent_containers/food/snacks/grown/cabbage/rogue, etc. All
of them were missing the /rogue on the end (except the beet, which does
not have such a variant).

## Why It's Good For The Game

Fixing things that are broken is good.

## Proof of Testing (Required)


![dreamseeker_2025-07-04_00-06-30](https://github.com/user-attachments/assets/1d035e19-1920-4049-834d-58b1d636e4e1)