Adds post_craft proc, offsets railing on craft (#2964)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
I originally wanted to make railings offset on craft to feel more like
they're in the proper place. It seems like the easiest way was to just
set a post_craft when they're made and call that. If there's a better
way let me know but this could be useful for other objects that need to
do something when they are crafted versus just initialized.

I might also consider making the south view be a rotated version of the
north sprite but left it unchanged for now for compatibility with
currently existing fences.

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Almost every railing on the map is designed with a pixel offset in mind.
When you craft them in game however you do not have this offset. I think
it looks more aesthetically pleasing to have the offset even in crafted
versions.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->


## Proof of Testing (Required)
3 railings on 1 tile before:

![image](https://github.com/user-attachments/assets/eed700d2-4bbd-4fdc-a1de-85c9f83f8d33)
3 railings on 1 tile after:

![image](https://github.com/user-attachments/assets/1c31162c-6d46-4315-8698-6f1b1c13b48d)

<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->