Removes tabaxi speech change (#3011)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
Removes the tabaxi handle_speech proc.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
People seem to hate it so it's here if people want to fight over if it
should exist or not.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->


## Proof of Testing (Required)

![image](https://github.com/user-attachments/assets/6440c93d-d851-4427-b582-a32b354ed35a)

<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->