Adds scrying orb failure effects, and reduces likelihood of the orb breaking (#3012)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
This PR adds a variety of failure effects that can occur when using the
scrying orb, based on the user's arcane skill and intelligence, and
whether or not the orb is on cooldown. These failure effects are
intended to be somewhat-punishing, but - mostly - more flavourful than
mechanically frustrating, with those with the harshest mechanical
penalties being the shortest-lived. I've also taken the opportunity to
refactor the orb code a little and fix some bugs (eg. the examine text
not working, and division by zero when a person with no arcane skill
attempts to use it).

Failure effects include:
- Scrying a random person other than the person you'd intended to scry.
- The hand holding the orb being burned by magical backlash.
- Being confused for a few minutes.
- Being put to sleep for a little over one minute.
- Being rendered drunk for a few minutes.
- Being rendered high for a few minutes.
- Having your intelligence and speed reduced, and having a "simple
speech" modifier applied.
- Being paralysed in a random selection of limbs for a couple of
minutes.
- Losing knowledge of Imperial for a few minutes.
- Becoming obsessed with the target of the scrying attempt, as if they'd
used a love potion on you.
- Having your intelligence drastically reduced and losing access to all
languages for 10 minutes (the worst effect - and very unlikely to be
seen in gameplay).

On failure to use the scrying orb, one of these failure effects will be
applied to the user depending on a few factors and a little bit of
randomisation. When a failure occurs, if the user has 2 or higher arcane
skill, the orb enters an extended ten-minute cooldown during which
success rate when using it is halved; if the user has 1 or 0, they still
receive failure effects as described above, but the orb breaks as it did
previously. Success rates while the orb is on cooldown have also been
reduced - in most cases, by about one skill level's worth, relative to
success rates when the orb is not on cooldown.

Finally - the "Simple speech" trait that was added as part of this is a
new speech modifier. Specifically, this speech modifier means a
character can only say the most common 1,000 English-language words
unhindered. Words not on this list have a chance of being replaced or
prefaced with "uh..." or "um...", and will be scrambled, leaving the
first and last letters in place to aid legibility while still
communicating the idea that the speaker is struggling with them.

Obviously, the specific "severities", durations, and so on of each
effect can be tweaked fairly easily if necessary.

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
The scrying orb is an interesting item with a lot of potential - but as
it stands, it's very risky to use if you aren't one of the roles that
spawn with it, due to the penalty for failing the skill check. With this
PR, failures would still be somewhat punishing - but wouldn't deprive
the user of the orb for the rest of the round, or until they're able to
find the funds to purchase a new one, thus making it a more worthwhile
investment for roles that have some magical capability.

The failures themselves, I think, are also much more interesting than
simply breaking the orb - often being effects that one can roleplay
around for, hopefully, a more generally interesting experience than
simply having to find money for a new orb.

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->


## Proof of Testing (Required)
I'll not show a fully comprehensive list of screenshots, but a few to
demonstrate certain elements working correctly. I promise I _have_
tested everything I can think of, though.

Drunkenness failure condition:

![image](https://github.com/user-attachments/assets/45ec79fe-bb3f-4eaa-883d-1e88ea2acfcd)

Dim-witted failure condition:

![image](https://github.com/user-attachments/assets/08f32949-be08-411a-b439-7e0cff8b10ab)

High failure condition:

![image](https://github.com/user-attachments/assets/2ff1ce46-9d8f-4705-b743-1e504ef4819a)

<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->