Vanderlin-Azure alchemy port [2] Meet new butching (#2993)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

Build butcherhook and pull the body to it to get +1 skill
Result depends of butching and random now
Changed drop from mobs so butchers and hunters are more valuable now
Will be more drop related to alchemy in future


## Why It's Good For The Game

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->


## Proof of Testing (Required)

<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->