Typo fix for 'river' tiles (#3055)
## About The Pull Request

Fixes a typo on the examine text for 'river' tiles. 
This PR is mostly serving as a way for me to figure out how GitHub even
works. I quite literally changed one line of code. If this somehow
breaks the entire server, I give anyone reading this full authority to
gun me down in the street.

## Why It's Good For The Game

One less typo! Huzzah!


## Proof of Testing (Required)

![Screenshot 2025-07-03
214210](https://github.com/user-attachments/assets/9fec4602-4683-4805-8298-d03eba055841)