Lowers boozepwr of Emberwine to reduce slurring. (#3058)
## About The Pull Request

Emberwine is a fun aphrodesiac with an enjoyable effect, good for
roleplay. Unfortunately, its Boozepwr is through the roof, which means
just a little bit is enough to get someone completely drunk off their
ass to the point where the game is making their speech extremely
difficult to understand and replacing large amounts of text with
burping. This PR lowers its "boozepwr" so that it no longer forces this.

## Why It's Good For The Game

Now, players who are chiefly interested in the aphrodesiac effect can
play around with that without having to put up with their speech being
massively slurred and replaced with *BURP*. If they DO like that, for
whatever reason, they can just drink a bunch of regular alcohol to go
along with their sex alcohol.

## Proof of Testing (Required)


![dreamseeker_2025-07-04_23-20-31](https://github.com/user-attachments/assets/734aa02a-4195-4f54-96be-ef967a680c2f)