AP Port - Key Summoning and Regency - - - Ports key summoning from AP. Ports regency and Steward tax setting from AP.
Regency might need to have the requirement to have no Priest, because it, as someone pointed out in that PR alone, makes them functionally useless for that purpose. That's why this is going into a draft, so someone can pick at it and figure something out, because I'm too lazy to care.