Rondel Daggers - - - Rondel daggers. Exactly as it says on the tin. Thin, pointed blades intended to punch into armour. Carried as a sidearm by knights, skilled warriors and the occasional brigand. Roles have been updated to reflect this.
Knight sentinels / champions can choose between the arming sword or bastard sword, respectively, or the rondel dagger and +1 to their knife skill.
The Captain gets one in his bag.
MAA Veterans get one.
Actual Veterans get one.
Hedge Knights get one.
Dagger Knaves lose one of their iron daggers, in exchange for a rondel dagger.
Merchants get one, because it was typically carried by tradesman and I felt it appropriate.

The weapon has been added to smithing. It's 2 steel, as opposed to the 1 of a normal steel dagger.
Smelting it gives one steel bar in return. Same as a steel dagger.
The damage is 14, which is 2 less than a combat knife, 1 less than a standard iron/steel dagger.
Weapon defence is 4, which is the highest of any daggers, with others having 3 typically.

Normally, the rondel dagger has a pick intent that only has 30% AP. Not great. Normal dagger stabbing has 40%, for reference. If you GRAB someone, however? It immediately becomes 80%, 10% higher than normal picking, should you pass a check determined by a combination of skill with the appropriate weapon (knives for the rondel dagger) and character speed. It has a slow enough attack on swift intent that responding isn't difficult, or so I'd hope. This'll need tweaking, if we even do add this.