Half-Orc NPCs (#3064)
## About The Pull Request
This is just the bum fights PR but with Half-Orcs.
That's it. I'm serious.
It doesn't introduce them into the map, but fixes old Orc NPCs up, makes
them a subspecies of Half-Orcs and changes paths to actually compile.
Also adjusts the loadouts some, but they'll need serious tweaking.

That's about it.

## Why It's Good For The Game
It doesn't currently touch anything player facing. Just readies them for
map use / events. This was a request by a mapper.
## Proof of Testing (Required)

![image](https://github.com/user-attachments/assets/5246aef1-6fe1-433c-85e4-65c9fb4fda09)