Alembic Quickfix & Magos Alchemy Raised (#3061)
## About The Pull Request
Changes the alembic to check for alchemy, as opposed to medicine. Also
provides the Magos one more skill in alchemy, so they can see what's
inside of bottles. I guess it helps with mass producing thermoscopes and
bottle bombs, but that's not really a concern.

That's it. That's the entire PR.

## Why It's Good For The Game
I really do believe the lack of ability to use their own alembic was
oversight, and the alembic is intended to be checking for alchemy as is.
Checking the contents of bottles is also something that I've seen the
Magos called upon for previously, but they'd been entirely incapable of
such. Which was pretty funny at the time, but not great as a player.
Physicians also exist for this purpose, and have the same alchemy skill,
to be quite clear. But they'd both be suitable for this task IMO.

## Proof of Testing (Required)
I didn't test this. It's a simple datum change and will crash /
repeatedly cause problems for the server.