Recipe book (#3041)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->
This adds a fully working recipe book, altrough I didnt add it being
spawned with people, there are already bookshelf subtypes that spawn all
recipe books in the game, for you to browse them. It opens a menu where
you will be able to see all craftable recipes, and as for blacksmithing,
you also will be able to see their icons and steps of production, along
with required skill level to craft anything.

## Why It's Good For The Game

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->
No more spending hours searching for correct recipe in the code, just to
craft a single silly belt.

## Proof of Testing (Required)

<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->

---------

Co-authored-by: WeNeedMorePhoron <132836028+WeNeedMorePhoron@users.noreply.github.com>
Co-authored-by: FreeStylaLT <freestyla8@gmail.com>