Clothing additions and changes from #2895 (#3043)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

Adds all the clothes that came from #2895 without touching the gambeson
which I believe is where the conflicts came from.
New sprites for the hood and volf helm

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

More clothes = Good! I love this PR.

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->


## Proof of Testing (Required)

<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->

![image](https://github.com/user-attachments/assets/38dbfb55-cd6b-48af-bbe9-52b1511303ec)

---------

Co-authored-by: Useroth <37159550+Useroth@users.noreply.github.com>