adds a tail + bug head sprite fix (#3047)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->
adds a new tail, and a BUG fix packaged alongside it, haha. bug fix.
## Why It's Good For The Game

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->
new tail + bug head was lacking a northern sprite, so it'd vanish
anytime you press W

## Proof of Testing (Required)

<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->

![image](https://github.com/user-attachments/assets/a1bf5eed-c65f-4901-a445-2446fd77698c)

![image](https://github.com/user-attachments/assets/2b81643e-eab6-479c-8752-24284464c51a)