Flaw Additions & Changes (#3052)
## About The Pull Request
This PR ports a number of flaws from AP, and otherwise adds a handful of
new, changed or returned features.
- - -
Ported:
 - Mute flaw. As you'd expect.
- Insomnia flaw. Disabled sleeping. So permanent debuff and no sleep
learning.
- Sadist flaw. Requires hearing screams within view. Fire, pain, or
agony.
- Various flaw inspects on examination. Meekness as a special hides
these altogether, just as it does on AP.
    - Nympho to Nympho: 'X as lovesick as I.'
- Junkie to Junkie: 'X carrying the same dust marks on their nose as I.'
- Smoker to Smoker: 'X enveloped by the familiar, faint stench of smoke.
I know it well.'
- Alcoholic to Alcoholic: 'X struggling to hide the hangover, and the
stench of spirits. We're alike.'
- Paranoid to Paranoid: 'X privy to the dangers of all these strangers
around us. X just as afraid as I am'
- Paranoid to Paranoid, other species: 'X one of the good ones. X just
as afraid as I am.'
- Sadist to Masochist: 'X marked by scars inflicted for pleasure. A
delectable target for my urges.'
- Masochist to Sadist: 'X looking with eyes filled with a desire to
inflict pain. So exciting.'
- - -
Re-enabled:
 - Nymphomaniac flaw. Also modifies recovery time.
- Devout flaw. Including changing the confession stress you get once
again.
- - -
New:
- Leper flaw. Removed from specials and made a flaw. Given it never
rolled, and when it did, most folks wouldn't ever touch it. Let us play
leper Dukes, I beg you.
- A few lines of different sated descriptors for addictions. Like a
sentence at most. Nothing big.
- - -
Other:
- Removes the Emberwine made re-addition of the Nymphomaniac flaw, given
it was a duplicate datum. It still functions as before, where your flaw
is swapped to it at stage 3 addiction.
- Ups the required length of prayers to receive benefits, including
sating the flaw, from 15, to 150.
- - -

## Why It's Good For The Game
Variety is the spice of life. Or something of the sort.
Removal of nymphomaniac was done given it encouraged some rather odd
behavior, but then, not long later, we get literal prostitute roles. I
don't know why you'd argue against this being returned, when we have
such content. Doubly so when we've events that would encourage such a
thing to exist.

Devout was too easy to sate, but it can be punished if it's abused by
putting in nonsense, without any issue on the part of staff. It
encourages RP and engaging with prayers. I can also increase the word
requirement if necessary, but I really don't see it being an issue.

The others speak for themselves. Sadist might be too easy of a choice
for certain roles as well, but given it has the examine interaction, and
is still far weaker of a choice than masochist, I'd argue it's fine.

## Proof of Testing (Required)

![image](https://github.com/user-attachments/assets/9a31110d-37e1-497c-9731-ed413bef0a39)

![image](https://github.com/user-attachments/assets/bd6025c4-8c74-4920-bcb0-02cf129bacd0)