Implements BYOS as a New Map (#3032)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

- Attempts to add BYOS into the map rotation, allowing it to be voted on
with a weight of 1 (Compared to 1.5 for Rockhill)
- Adds changes to the BYOS map to make it less lethal, and more of a
normal gameplay map. Such changes include:
 - Hoardmaster and furnaces for bandits 
 - Mini Vampire Manor
 - Independent merchant
 - Merchant ship
 - Small Prison in the ship
 - Removed bosses and reduced mob count
 - Removes the abysmal slowdown effect for the sand tile
 

I'm not a mapper, nor do I know if this is implemented correctly. I will
definitely need some help with this one, hence why it's starting as a
Draft.

## Why It's Good For The Game

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->

The BYOS is different enough to warrant it's return, and I believe it's
a good basis for a default map going forward, especially for lowpop
rounds with few antagonists. It has a lot of dungeons and PvE content,
and will garner a lot more interaction between players, forcing them to
work together to build a new settlement.

The map has a lot of space for more to be added to it (A good portion of
a certain z-layer is blank). I believe with good upkeep, the
understanding that it's in development, this map could become a decent
main-stay for the server. The work Mario and Opaqh did on it feels like
a waste to only be used once with how great it is.

## Proof of Testing (Required)

<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->

![image](https://github.com/user-attachments/assets/fa54c5b2-c481-466a-9252-48ead83442c1)

![image](https://github.com/user-attachments/assets/40a9970e-e43d-4312-801c-b5c9184c437f)

![image](https://github.com/user-attachments/assets/5875ef19-1609-4ef9-86c9-6b528a781dbd)

---------

Co-authored-by: Useroth <37159550+Useroth@users.noreply.github.com>