Rondel Daggers & Misericordes (#3071)
## About The Pull Request
Rondel daggers and misericordes. Exactly as it says on the tin. Thin,
pointed blades intended to punch into armour. Carried as a sidearm by
knights, skilled warriors and the occasional brigand. Roles have been
updated to reflect this.

Knight sentinels / champions can choose between the arming sword or
bastard sword, respectively, or the rondel dagger and +1 to their knife
skill. The Captain gets a rondel dagger in his bag.
MAA Veterans get a misericorde.
Actual Veterans get a misericorde.

Hedge Knights get a rondel dagger.
Dagger Knaves lose one of their iron daggers, in exchange for a
misericorde.

Merchants get a rondel dagger, because it was typically carried by
tradesman and I felt it appropriate.

All mercenary sellsword daggers have been replaced with a misericorde.
New Moon is provided a misericorde instead of an iron dagger.
Grenzelhoft Jaegers have their hunting knife replaced by a misericorde.

Both weapons have been added to smithing. It's 2 steel, as opposed to
the 1 of a normal steel dagger for the rondel dagger. Identical but with
iron for the misericorde. Smelting the rondel gives one steel bar in
return. Same as a steel dagger. Identical again with the misericorde,
but for iron.
The damage on the rondel is 14, which is 2 less than a combat knife, 1
less than a standard iron/steel dagger. Weapon defence is 4, which is
the highest of any daggers, with others having 3 typically.
The misericorde as an iron cousin of the rondel has 13 damage, 1
defence, and less durability, while also missing the cut intent. It
makes up for this by being slightly faster in its strikes.
Both daggers are capable of mining as a consequence of using the pick
bladeclass. I determined this fine, because they're utility items
primarily. This also differentiates them from actual daggers.

Unique to the special grapple combat this dagger allows, is that they'll
swap to a no penetration smash intent if you target the ears. This is to
prevent grapple two-tap ear stabs. You can let them go instead of
grappling to bypass this, though you'll lose the penetration in the
process.

Normally, the two daggers have a pick intent that only has 30% AP. Not
great. Normal dagger stabbing has 40%, for reference. If you GRAB
someone, however? It immediately becomes 80%, 10% higher than normal
picking, should you pass a check determined by a combination of skill
with the appropriate weapon (knives for the rondel dagger) and character
speed. It has a slow enough attack on swift intent that responding isn't
difficult, or so I'd hope. This'll need tweaking, if we even do add
this.

## Why It's Good For The Game
Dagger users either have to rely on a ten thousand year windup to deal
with armour, or fall back on other methods instead. This should
alleviate that, while also giving speed based knife users a way to deal
with armour that isn't just cheese. Atop that, a dagger that can defend
quite well.

Additionally, this is more thematic than a bastard sword in the case of
knight champions, which they can still choose, but it'd be weird and
redundant, given the big sword you get anyways.

They're also time appropriate.
## Proof of Testing (Required)
<img width="181" height="84" alt="image"
src="https://github.com/user-attachments/assets/b9bfe4d4-99dd-43d9-b5bf-eea888e04df4"
/>

Bashed together sprites from a few different oldsprites above.
Tested locally. Forgot to take screenshots.