Cicerone trait from AP + Eargrab trait (#3069)
## About The Pull Request

Just adds the Cicerone trait from AP, which lets you view reagents. I
plan on giving it to the Head Butler, the Innkeep, and the Merchant. It
also adds an "eargrab" trait, which is identical to nosegrab, but lets
you grab ears. I plan on giving this to the Priest and the Consort.

## Why It's Good For The Game

It makes sense for a butler, innkeep, and merchant to be able to
identify different alcohols. I'm surprised it took this long to port it,
honestly.
A priest grabbing people by the ears just fits thematically. Same with
the consort.

## Proof of Testing (Required)


![image](https://github.com/user-attachments/assets/2e948c0f-b9e2-4996-9707-68847dbc9ccd)

![image](https://github.com/user-attachments/assets/d53e8087-e8fa-42d9-85e2-4f3c0420f726)

![image](https://github.com/user-attachments/assets/4c8cb518-8145-4cb7-96d2-f40c16c63c89)